% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshmetrics.r
\name{ang}
\alias{ang}
\title{Cosine triangle function to calculate angle C (in degrees)
given edge lengths a, b, and c}
\usage{
ang(a, b, c)
}
\arguments{
\item{a}{A numeric triangle edgelength (opposite angle A).}

\item{b}{A numeric triangle edgelength (opposite angle B).}

\item{c}{A numeric triangle edgelength (opposite angle C).}
}
\description{
Cosine triangle function to calculate angle C (in degrees)
given edge lengths a, b, and c
}
\examples{
\dontrun{
ang(1, 2, 1.5)
}
}
