% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{bigfoot}
\alias{bigfoot}
\title{Bigfoot (Sasquatch) sightings in the USA, 2000 - 2005}
\format{
A simple features dataframe, of type \code{POINT}, with 972 observations and 27 fields:
\describe{
\item{observed}{Text observation summary}
\item{location_details}{Text location summary}
\item{county}{County}
\item{state}{State}
\item{season}{Season}
\item{title}{Report title}
\item{date}{Date}
\item{number}{Report number}
\item{classification}{Report classification}
\item{geohash}{Geohash code}
\item{temperature_high}{Repotred weather measure}
\item{temperature_mid}{Repotred weather measure}
\item{temperature_low}{Repotred weather measure}
\item{dew_point}{Repotred weather measure}
\item{humidity}{Repotred weather measure}
\item{cloud_cover}{Repotred weather measure}
\item{moon_phase}{Repotred measure}
\item{precip_intensity}{Repotred weather measure}
\item{precip_probability}{Repotred weather measure}
\item{precip_type}{Repotred weather measure}
\item{pressure}{Repotred weather measure}
\item{summary}{Text weather summary}
\item{uv_index}{Repotred weather measure}
\item{visibility}{Repotred weather measure}
\item{wind_bearing}{Repotred weather measure}
\item{wind_speed}{Repotred weather measure}
\item{year}{Year}
}
}
\source{
\url{https://data.world/timothyrenner/bfro-sightings-data}
}
\description{
Subset of data sourced from the Bigfoot Field Researchers Organization (BFRO)
(\url{https://data.world/timothyrenner/bfro-sightings-data}).
}
