% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshmetrics.r
\name{circum_O}
\alias{circum_O}
\title{Calculate cicumcircle centroid given vertex A, B, C
locations of a triangle}
\usage{
circum_O(A, B, C)
}
\arguments{
\item{A}{A numeric vector of length 2 specifying vertex location "A".}

\item{B}{A numeric vector of length 2 specifying vertex location "B".}

\item{C}{A numeric vector of length 2 specifying vertex location "C".}
}
\description{
Calculate cicumcircle centroid given vertex A, B, C
locations of a triangle
}
\examples{
\dontrun{
circum_O(c(1, 1), c(2, 1), c(3, 2.5))
}
}
