% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{coef.stelfi_tmb}
\alias{coef.stelfi_tmb}
\title{Extract estimated coefficients from an object of class \code{spde_tmb}}
\usage{
\method{coef}{stelfi_tmb}(obj, se = FALSE)
}
\arguments{
\item{se}{Logical, default FALSE. If TRUE then standard errors are also returned.}
}
\description{
Extract estimated coefficients from an object of class \code{spde_tmb}
}
