% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{compensator_differences}
\alias{compensator_differences}
\title{Extract the compensator differences}
\usage{
compensator_differences(obj)
}
\description{
Extract the compensator differences from a fitted Hawkes
model.
}
\seealso{
\code{\link{show_hawkes_GOF}}
}
