% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.r
\name{compile.stelfi}
\alias{compile.stelfi}
\title{Fuction to compile TMB C++ templates}
\usage{
compile.stelfi(x = NULL)
}
\description{
Compiles the  TMB templates into a shared object file.
Must be done a single time following installation or updating of the package.
}
