% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshmetrics.r
\name{dist}
\alias{dist}
\title{Calculate the distance between two points}
\usage{
dist(A, B)
}
\arguments{
\item{A}{A numeric vector of length 2 specifying (first) location.}

\item{B}{A numeric vector of length 2 specifying (second) location.}
}
\description{
Calculate the distance between two points
}
\examples{
\dontrun{
dist(c(1, 1), c(2, 1))
}
}
