% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.r
\name{dll.stelfi}
\alias{dll.stelfi}
\title{Function to load DLLs for C++ templtes
Loads required DLLs for models fitted using TMB}
\usage{
dll.stelfi(x = NULL)
}
\description{
Function to load DLLs for C++ templtes
Loads required DLLs for models fitted using TMB
}
