% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.r
\name{dll_stelfi}
\alias{dll_stelfi}
\title{Function to load DLLs for C++ templates}
\usage{
dll_stelfi(x)
}
\arguments{
\item{x}{Optional, if provided specifies which \code{\link{stelfi}} DLL to load.}
}
\description{
\code{dll_stelfi} loads required DLLs for models fitted using TMB.
For compiling the required templates see \code{\link{compile_stelfi}}.
}
\examples{
\dontrun{
dll_stelfi()
}
}
