% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{dual_mesh}
\alias{dual_mesh}
\title{Internal function to construct the `dual` mesh}
\source{
\url{http://www.r-inla.org/spde-book}
}
\usage{
dual_mesh(mesh)
}
\arguments{
\item{mesh}{A spatial mesh of class \code{\link[INLA]{inla.mesh.2d}}.}
}
\value{
An \code{sf} object of the Voronoi tessellation
centered at each \code{mesh} node.
}
\description{
Internal function to construct the `dual` mesh
}
\seealso{
\url{http://www.r-inla.org/spde-book} and \url{https://becarioprecario.bitbucket.io/spde-gitbook/}
}
