% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lgcp.r
\name{fit.lgcp}
\alias{fit.lgcp}
\title{Function to fit a spatiotemporal log-Gaussian Cox process using TMB and the
R_inla namespace for the spde construction of the latent field}
\usage{
fit.lgcp(locs, temp.idx, mesh, parameters, sp, tmb = TRUE, ...)
}
\arguments{
\item{locs}{a matrix of locations 2xn}

\item{temp.idx}{numeric aggregated temporal index 1,...,n}

\item{mesh}{a Delaney triangulation of the study area created using \code{INLA::inla.mesh.2d()}}

\item{parameters}{a named vector of parameters. Must include "beta" the regression coefficients of fixed effects,
"log_kappa" the smoothness parameter of the random field, if an AR(1) process for time is being fitted then
"rho" must also be included as the temporal dependence parameter}

\item{sp}{optional spatial polygon of the domain. Should be supplied for lgcp as spde
should set weigths to 0 outside of domain}

\item{tmb}{logical if TRUE will use \code{TMB} framework to fit model (default), otherwise \code{INLA}}

\item{...}{arguments to pass into \code{nlminb()} if using TMB, if using INLA framework arguments passed into
\code{inla()}}

\item{covs}{a}
}
\description{
Function to fit a spatiotemporal log-Gaussian Cox process using TMB and the
R_inla namespace for the spde construction of the latent field
}
