% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hawkes.r
\name{fit_hawkes_cbf}
\alias{fit_hawkes_cbf}
\title{Function to fit a self-exciting Hawkes process with custom background function (cbf) using TMB
The \code{alpha} and \code{beta} parameters are estimated using TMB,
parameters of the CBF are optimized in R}
\usage{
fit_hawkes_cbf(
  times,
  parameters,
  model = 1,
  marks = c(rep(1, length(times))),
  background,
  background_integral,
  background_parameters,
  background_min,
  tmb_silent = TRUE,
  optim_silent = TRUE
)
}
\arguments{
\item{times}{a vector of numeric observation time points}

\item{parameters}{a vector of named parmeters for the chosen model (see \code{model})
"mu"--base rate of the hawkes process,
"alpha"--intensity jump after an event occurence, and
"beta"--exponential intensity decay}

\item{model}{a factor indicator specifying which model to fit:
\code{1}, a hawkes process with exponential decay and cbf (default);
\code{2}, a hawkes process and cbf with negative self-exciting intensity "jump".}

\item{marks}{a vector of numerical marks, defaults to 1 (i.e., no marks)}

\item{background}{a function taking one parameter and an independent variable, returning a scalar}

\item{background_integral}{the integral of background}

\item{background_min}{a function taking one parameter and two points,
returns min of background between those points}

\item{tmb_silent}{logical, default `TRUE`:
TMB inner optimization tracing information will be printed.}

\item{optim_silent}{logical, default `TRUE`:
print function and parameters every iteration.}

\item{background_parameter}{the parameter (which could be a list of multiple values)
for the background function}
}
\description{
Function to fit a self-exciting Hawkes process with custom background function (cbf) using TMB
The \code{alpha} and \code{beta} parameters are estimated using TMB,
parameters of the CBF are optimized in R
}
