% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stelfi.r
\name{fit_hspde_tmb}
\alias{fit_hspde_tmb}
\title{Function to fit a spde spatial hawkes process using \code{TMB} and the
\code{R_inla} namespace for the spde construction of the latent field.}
\usage{
fit_hspde_tmb(
  times,
  locs,
  sp,
  mesh,
  lmat,
  log_mu = 0,
  logit_abratio = 0,
  log_beta = 0,
  log_kappa = 0,
  log_tau = 0,
  log_xsigma = 0,
  log_ysigma = 0,
  atanh_rho = 0,
  spde,
  w,
  tmax = max(times),
  reltol = 1e-12,
  abstol = 1e-12,
  tmb_silent,
  nlminb_silent,
  ...
)
}
\arguments{
\item{times}{a vector of numeric observation time points}

\item{locs}{2xn \code{data.frame} of locations x, y. If locations have
time stapms then this is the third column of the 3xn matrix}

\item{sp}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
of the domain}

\item{tmb_silent}{logical, default `TRUE`:
TMB inner optimization tracing information will be printed.}

\item{nlminb_silent}{logical, default `TRUE`:
print function and parameters every iteration.}

\item{...}{arguments to pass into \code{optim()}}
}
\description{
Function to fit a spde spatial hawkes process using \code{TMB} and the
\code{R_inla} namespace for the spde construction of the latent field.
}
