% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lgcp.r
\name{fit_lgcp}
\alias{fit_lgcp}
\title{Function to fit a spatial or spatiotemporal log-Gaussian Cox process using \code{TMB}}
\usage{
fit_lgcp(locs, sp, smesh, tmesh, beta, tau, kappa, rho, silent, ...)
}
\arguments{
\item{locs}{2xn \code{data.frame} of locations x, y. If locations have
time stapms then this is the third column of the 3xn matrix}

\item{sp}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} of the domain}

\item{smesh}{spatial mesh of class \code{"inla.mesh"}}

\item{tmesh}{optional, temporal mesh of class \code{"inla.mesh.1d"}}

\item{beta}{A vector of fixed effects needs to be estimated
(same length as \code{nrow(designmat)}}

\item{tau}{tau parameter for the GMRF}

\item{kappa}{kappa parameter for the GMRF}

\item{rho}{optional, rho AR1 parameter}

\item{silent}{logical, by default FALSE. If TRUE model fitting progress
not printed to console.}

\item{...}{arguments to pass into \code{nlminb()}}
}
\description{
A simple to use wrapper for \code{fit_lgcp_tmb}
}
