% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_coefs}
\alias{get_coefs}
\title{Reported parameter estimates}
\usage{
get_coefs(object)
}
\arguments{
\item{object}{result of a call to \code{fit_hawkes()} or \code{fit_lgcp()}}
}
\description{
Reported parameter estimates
}
