% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_fields}
\alias{get_fields}
\title{Estimated random field(s)}
\usage{
get_fields(object, smesh, tmesh, plot = FALSE, sd = FALSE)
}
\arguments{
\item{object}{the result of a call to \code{\link{fit_lgcp}}.}

\item{smesh}{A spatial mesh of class \code{\link[INLA]{inla.mesh.2d}}.}

\item{tmesh}{Optional, a temporal mesh of class \code{\link[INLA]{inla.mesh.1d}}.}

\item{plot}{\code{logical}, if \code{TRUE} then the returned values are plotted.}

\item{sd}{\code{logical}, if \code{TRUE} then standard errors of field are returned.}
}
\description{
\code{get_fields} returns
}
