% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{hawke.intensity}
\alias{hawke.intensity}
\title{Hawkes intensty function with decay historical dependence}
\usage{
hawke.intensity(mu, alpha, beta, times, p = NULL)
}
\arguments{
\item{mu}{numeric integer specifying base rate of the hawkes process}

\item{alpha}{numeric integer specifying intensity jump after an event occurence}

\item{beta}{numeric integer specifying exponential intensity decay}
}
\description{
Hawkes intensty function with decay historical dependence
}
