% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{hawkes_intensity}
\alias{hawkes_intensity}
\title{Hawkes intensity function}
\usage{
hawkes_intensity(
  times,
  mu,
  alpha,
  beta,
  p,
  marks = rep(1, length(times)),
  background_parameters
)
}
\arguments{
\item{times}{A vector of numeric observed time points.}

\item{mu}{A numeric specifying base rate of the Hawkes process.}

\item{alpha}{A numeric specifying intensity jump after an event
occurrence.}

\item{beta}{A numeric specifying exponential intensity decay}

\item{p}{An optional vector of pseudo times at which to calculate the intensity.}

\item{marks}{Optional, a vector of numeric marks, defaults to 1 (i.e., no marks).}

\item{background_parameters}{The parameter(s)
for the background function \code{background}.
This could be a list of multiple values.}
}
\description{
Calculate the Hawkes intensity given a vector
of time points and parameter values
}
\examples{
times <- sim_hawkes(10.2, 3.1, 8.9, method = "2")
hawkes_intensity(times = times, 10.2, 3.1, 8.9)

}
