% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_wrappers.r
\name{inla.mesh.dual}
\alias{inla.mesh.dual}
\title{Function to construct the dual mesh}
\source{
\url{http://www.r-inla.org/spde-book}
}
\usage{
inla.mesh.dual(mesh)
}
\description{
\code{inla.mesh.dual} is an internal function that returns a list of \code{SpatialPolygons}
centered at each \code{mesh} node.
}
\seealso{
\url{http://www.r-inla.org/spde-book} and \url{https://becarioprecario.bitbucket.io/spde-gitbook/}
}
