% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{iraq_terrorism}
\alias{iraq_terrorism}
\title{Terrorism in Iraq, 2013 - 2017}
\format{
A simple features dataframe, of type \code{POINT}, with 4208 observations and 16 fields:
\describe{
\item{iyear}{numeric year 2013--2017}
\item{imonth}{numeric month index 1--12}
\item{iday}{numeric day 1--31 (zeros are a non-entry)}
\item{country}{country (IRAQ)}
\item{latitude}{latitude location}
\item{longitude}{longitude location}
\item{utm_x}{x-coord location UTM}
\item{utm_y}{y-coord location UTM}
\item{success}{logical, was fatal? \code{TRUE} = fatal}
\item{nkill}{number of fatalities per attack}
\item{specificity}{spatial accuracy of event: 1 = most accurate, 5 = worst}
\item{gname}{character name of attack perpetrators (ISIL)}
\item{x_coord}{x coordinate from location projected onto a sphere}
\item{y_coord}{y coordinate from location projected onto a sphere}
\item{z_coord}{z coordinate from location projected onto a sphere}
\item{popdensity}{scaled: number of people per kilometer squared}
\item{luminosity}{scaled: luminosity}
\item{tt}{scaled: time to nearest city in minutes}
}
}
\source{
\url{https://www.start.umd.edu/gtd/}
}
\description{
A dataset containing information of terrorism activity carried out
by the Islamic State of Iraq and the Levant (ISIL) in Iraq,
2013 - 2017.
}
