% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshmetrics.r
\name{lmin}
\alias{lmin}
\title{Calculate minimum edge length for each triangle in a given Delaunay triangulation}
\usage{
lmin(mesh)
}
\arguments{
\item{mesh}{A \code{\link[INLA]{inla.mesh.2d}} object.}
}
\description{
Calculate minimum edge length for each triangle in a given Delaunay triangulation
}
\examples{
\dontrun{
data(horse_mesh, package = "stelfi")
lmin(horse_mesh)
}
}
