% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{make.covs}
\alias{make.covs}
\title{Function that takes in a named  matrix of covariates with
ncol equal to the number of covariates
returns a list containing the effects ready to be
read by \code{inla.stack} and a covariate formula,
ready to be read by a call to \code{fit_lgcp_inla}}
\usage{
make.covs(covariates)
}
\description{
Function that takes in a named  matrix of covariates with
ncol equal to the number of covariates
returns a list containing the effects ready to be
read by \code{inla.stack} and a covariate formula,
ready to be read by a call to \code{fit_lgcp_inla}
}
