% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{marked}
\alias{marked}
\title{Example marked point pattern data set}
\format{
A data frame with 159 rows and 5 variables:
\describe{
\item{x}{x coordinate}
\item{y}{y coordinate}
\item{m1}{mark, Gaussian distributed}
\item{m2}{mark, Bernoulli distributed}
\item{m3}{mark, Gamma distributed}
}
}
\description{
Example marked point pattern data set
}
