% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshmetrics.r
\name{mesh_2_sf}
\alias{mesh_2_sf}
\title{Transform a \code{\link[INLA]{inla.mesh.2d}} into a \code{\link[sf]{sf}}}
\source{
\url{https://groups.google.com/g/r-inla-discussion-group/c/z1n1exlZrKM}
}
\usage{
mesh_2_sf(mesh)
}
\arguments{
\item{mesh}{A \code{\link[INLA]{inla.mesh.2d}} object.}
}
\value{
A \code{\link[sf]{st_as_sf}}.
}
\description{
Transform a \code{\link[INLA]{inla.mesh.2d}} into a \code{\link[sf]{sf}}
}
\details{
Modified from function suggested by Finn in the
R-inla discussion Google Group \url{https://groups.google.com/g/r-inla-discussion-group/c/z1n1exlZrKM}.
}
\examples{
\dontrun{
data(horse_mesh, package = "stelfi")
mesh_2_sf(horse_mesh)
}
}
