% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lgcp.r
\docType{class}
\name{missing_or_spatialpolygon-class}
\alias{missing_or_spatialpolygon-class}
\title{creates a virtual class that is a superclass to the component classes so then both
children inherit from that class}
\description{
creates a virtual class that is a superclass to the component classes so then both
children inherit from that class
}
