% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{multi_hawkes}
\alias{multi_hawkes}
\title{Example multivariate Hawkes dataset}
\format{
A data frame with 213 observations and 2 variables:
\describe{
\item{times}{ordered time stamp of observation}
\item{stream}{character giving stream ID (i.e., Stream 1 or
Stream 2) of observation}
}
}
\description{
Two-stream multivariate Hawkes data with \eqn{mu_1 = \mu_2 = 0.2},
\eqn{beta_1 = \beta_2 = 0.7}, and \eqn{\boldsymbol{\alpha}} =
\code{matrix(c(0.5,0.1,0.1,0.5),byrow = TRUE,nrow = 2)}.
}
