% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{murdersNZ}
\alias{murdersNZ}
\title{A dataset of recorded murder cases in New Zealand 2004--2019
scraped from the online database https://interactives.stuff.co.nz/2019/the-homicide-report/}
\format{A dataframe with 1019 rows and 9 variables
\describe{
\item{Latitude}{approximate latitude of murder}
\item{Longitude}{approximate longitude of murder}
\item{Sex}{biological sex of victim}
\item{Age}{age of victim (years)}
\item{Date}{month and day of murder}
\item{Year}{year}
\item{Cause}{cause of death}
\item{Killer}{killer}
\item{Name}{name of victim}
}}
\usage{
data(murdersNZ)
}
\description{
A dataset of recorded murder cases in New Zealand 2004--2019
scraped from the online database https://interactives.stuff.co.nz/2019/the-homicide-report/
}
