% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\docType{class}
\name{numericOrNULL-class}
\alias{numericOrNULL-class}
\title{creates a virtual class that is a superclass to the component classes so then both children inherit from that class}
\description{
creates a virtual class that is a superclass to the component classes so then both children inherit from that class
}
