% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{nz_earthquakes}
\alias{nz_earthquakes}
\title{Earthquakes in Canterbury, NZ, 2010 - 2014}
\format{
A simple features dataframe, of type \code{POINT}, with 3824 observations and 3 fields:
\describe{
\item{origintime}{The UTC time of the event's occurrence}
\item{magnitude}{The magnitude of the earthquake}
\item{depth}{The focal depth of the event (km)}
}
}
\source{
\url{http://quakesearch.geonet.org.nz/}
}
\description{
Earthquake data from Canterbury,
New Zealand 16-Jan-2010--24-Dec-2014.
}
