% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{nz_murders}
\alias{nz_murders}
\title{Murders of NZ, 2004 - 2019}
\format{
A simple features dataframe, of type \code{POINT}, with 967 observations and 11 fields:
\describe{
\item{sex}{Biological sex of victim}
\item{age}{Age of victim (years)}
\item{date}{Month and day of murder}
\item{year}{Year}
\item{cause}{Cause of death}
\item{killer}{Killer}
\item{name}{Name of victim}
\item{full_date}{Date object of observation on single days}
\item{month}{Month name of observation}
\item{cause_cat}{Cause of death as category}
\item{region}{NZ region}
}
}
\source{
Data scraped and cleaned by Charlie Timmings,
honours student at the University of Auckland from the website
\url{https://interactives.stuff.co.nz/2019/the-homicide-report/}
}
\description{
A dataset of recorded murder cases in New Zealand 2004 - 2019.
}
