% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hawkes.r, R/helpers.r
\name{outwith}
\alias{outwith}
\title{Function to find mesh nodes outwith some spatial polygon
relies on inla.mesh.dual function from INLA spde-tutorial}
\usage{
outwith(mesh = NULL, boundary = NULL)

outwith(mesh = NULL, boundary = NULL)
}
\description{
Function to find mesh nodes outwith some spatial polygon
relies on inla.mesh.dual function from INLA spde-tutorial

Function to find mesh nodes outwith some spatial polygon
relies on inla.mesh.dual function from INLA spde-tutorial
}
