% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{owin_to_polygons}
\alias{owin_to_polygons}
\title{\url{https://stat.ethz.ch/pipermail/r-sig-geo/2009-May/005781.html}}
\usage{
owin_to_polygons(x, id = "1")
}
\description{
\url{https://stat.ethz.ch/pipermail/r-sig-geo/2009-May/005781.html}
}
