% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{owin_to_sp}
\alias{owin_to_sp}
\title{Function to convert a\code{spatstat} \code{owin} object to
a \code{SpatialPolygonsDataFrame}}
\source{
\url{https://stat.ethz.ch/pipermail/r-sig-geo/2009-May/005781.html}
}
\usage{
owin_to_sp(x)
}
\arguments{
\item{x}{an object of class \code{owin}}
}
\value{
a \code{SpatialPolygonsDataFrame}
}
\description{
Function to convert a\code{spatstat} \code{owin} object to
a \code{SpatialPolygonsDataFrame}
}
