% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshmetrics.r
\name{plot_mesh}
\alias{plot_mesh}
\title{Plot a \code{\link[INLA]{inla.mesh.2d}} object by attribute}
\usage{
plot_mesh(mesh, radius_edge = TRUE)
}
\arguments{
\item{mesh}{A \code{\link[INLA]{inla.mesh.2d}} object.}

\item{radius-edge}{Logical, if \code{TRUE} (default) then the radius-edge ratio
of each triancle in the Delaunay triangulation is plotted. If \code{FALSE}
then the radius-ratio is plotted.}
}
\description{
Plot a \code{\link[INLA]{inla.mesh.2d}} object by attribute
}
\examples{
\dontrun{
data(horse_mesh, package = "stelfi")
plot_mesh(horse_mesh)
plot_mesh(horse_mesh,  radius_edge = FALSE)
}
}
\seealso{
\code{\link{meshmetrics}}
}
