% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{points.in.mesh}
\alias{points.in.mesh}
\title{Function that takes in a list of points 
(and optionally a weight or covariate for each point)
and a mesh of polygons
Returns either the number of points in each polygon or sum of the weights}
\usage{
\method{points}{in.mesh}(xy, dmesh, weights)
}
\description{
Function that takes in a list of points 
(and optionally a weight or covariate for each point)
and a mesh of polygons
Returns either the number of points in each polygon or sum of the weights
}
