% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lgcp.r
\name{prep}
\alias{prep}
\title{Function to prep INLA stuff}
\usage{
prep(locs, sp, smesh, tmesh)
}
\arguments{
\item{locs}{2xn \code{data.frame} of locations x, y. If locations have
time stapms then this is the third column of the 3xn matrix}

\item{sp}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} of the domain}

\item{smesh}{spatial mesh of class \code{"inla.mesh"}}

\item{tmesh}{optional, temporal mesh of class \code{"inla.mesh.1d"}}
}
\description{
Function to prep INLA stuff
}
