% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{retweets_niwa}
\alias{retweets_niwa}
\title{Retweets of NIWA's viral leopard seal Tweet}
\format{
A vector of length 4890 specifying the date and time of retweet (UTC)
}
\source{
\url{https://twitter.com/niwa_nz/status/1092610541401587712}
}
\description{
A dataset of retweet times of NIWA's viral leopard seal tweet
on the 5th Feb 2019 (\url{https://twitter.com/niwa_nz/status/1092610541401587712}).
}
