% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{retweetsNIWA}
\alias{retweetsNIWA}
\title{A dataset of retweets of NIWA's "viral" leopard seal tweet https://twitter.com/niwa_nz/status/1092610541401587712
on the 5th Feb 2019}
\format{A vector of length 4890
\describe{
\item{}{date and time of retweet}
}}
\description{
A dataset of retweets of NIWA's "viral" leopard seal tweet https://twitter.com/niwa_nz/status/1092610541401587712
on the 5th Feb 2019
}
