% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{sasquatch}
\alias{sasquatch}
\title{Sasquatch (bigfoot) sightings in the USA, 2000 - 2005}
\format{
A simple features dataframe, of type \code{POINT}, with 972 observations and 27 fields:
\describe{
\item{observed}{Text observation summary}
\item{location_details}{Text location summary}
\item{county}{County}
\item{state}{State}
\item{season}{Season}
\item{title}{Report title}
\item{date}{Date}
\item{number}{Report number}
\item{classification}{Report classification}
\item{geohash}{Geohash code}
\item{temperature_high}{Reported weather measure}
\item{temperature_mid}{Reported weather measure}
\item{temperature_low}{Reported weather measure}
\item{dew_point}{Reported weather measure}
\item{humidity}{Reported weather measure}
\item{cloud_cover}{Reported weather measure}
\item{moon_phase}{Reported measure}
\item{precip_intensity}{Reported weather measure}
\item{precip_probability}{Reported weather measure}
\item{precip_type}{Reported weather measure}
\item{pressure}{Reported weather measure}
\item{summary}{Text weather summary}
\item{uv_index}{Reported weather measure}
\item{visibility}{Reported weather measure}
\item{wind_bearing}{Reported weather measure}
\item{wind_speed}{Reported weather measure}
\item{year}{Year}
}
}
\source{
\url{https://data.world/timothyrenner/bfro-sightings-data}
}
\description{
Subset of data sourced from the Bigfoot Field Researchers Organization (BFRO)
(\url{https://data.world/timothyrenner/bfro-sightings-data}).
}
