% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshmetrics.r
\name{segments}
\alias{segments}
\title{Extract a dataframe of mesh triangle segments (start and end locations)}
\usage{
segments(mesh)
}
\arguments{
\item{mesh}{A \code{\link[INLA]{inla.mesh.2d}} object.}
}
\description{
Extract a dataframe of mesh triangle segments (start and end locations)
}
\examples{
\dontrun{
data(horse_mesh, package = "stelfi")
segments(horse_mesh)
}
}
