% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{serialUK}
\alias{serialUK}
\title{A dataset of recorded serial killers in the UK
constructed from http://www.murderuk.com/}
\format{A dataframe with 62 rows and 8 variables
\describe{
\item {Number.of.kills}{approximate number of murders committed by each person}
\item {Years}{The years of operation}
\item {Name}{Name of convicted serial killer}
\item {AKA}{Some serial killers were given nicknames}
\item {Year.Start}{Year the murders began}
\item {Year.End}{Year the murders ended}
\item {Date.of.first.kill}{Date, if known, first murder was committed}
\item {population..M.}{Estimated population of UK in millions at time of first murder}
}}
\usage{
data(serialUK)
}
\description{
A dataset of recorded serial killers in the UK
constructed from http://www.murderuk.com/
}
