% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\docType{methods}
\name{show.field}
\alias{show.field}
\title{Function to plot or create raster of a given field defined over a mesh}
\usage{
show.field(x, mesh, dims = c(300, 300), col = terrain.colors(100),
  sp = NULL, rast = FALSE, legend = TRUE, legend.only = FALSE, ...)
}
\arguments{
\item{x}{vector of length \link{mesh$n} of values at each \link{mesh} node}

\item{mesh}{a Delaney triangulation of the study area created using \code{INLA::inla.mesh.2d()}}

\item{dims}{vector of length 2 defining how fine a projection; default c(300,300)}

\item{col}{colours of plot; default terrain.colors(100)}

\item{sp}{optional spatial polygon of the domain
should set weigths to 0 outside of domain}

\item{rast}{logical; if TRUE create raster object; default FALSE}

\item{legend}{logical; plot legend; default TRUE}

\item{legend.only}{logical; legend only to be plotted; default FALSE}

\item{...}{arguments to pass into raster (if rast = TRUE) or image()}
}
\description{
Function to plot or create raster of a given field defined over a mesh
}
