% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\docType{methods}
\name{show_field}
\alias{show_field}
\title{Plots estimated random field(s) of a LGCP}
\usage{
show_field(x, smesh, dims = c(500, 500))
}
\arguments{
\item{x}{a vector of values for each mesh node}

\item{smesh}{spatial mesh of class \code{"inla.mesh"}}

\item{dims}{by default \code{c(500,500)}, vector of length 2 specifying
spatial pixel resolution}
}
\description{
Plots estimated random field(s) of a LGCP
}
