% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{show_hawkes_GOF}
\alias{show_hawkes_GOF}
\title{Plot the compensator ans other associated goodness of fit metrics for a Hawkes process}
\usage{
show_hawkes_GOF(
  times,
  mu,
  alpha,
  beta,
  marks = c(rep(1, length(times))),
  background_param,
  plot = TRUE,
  return_values = TRUE
)
}
\arguments{
\item{times}{a vector of numeric observation time points}

\item{mu}{numeric integer specifying base rate of the hawkes process}

\item{alpha}{numeric integer specifying intensity jump after an event
occurence}

\item{beta}{numeric integer specifying exponential intensity decay}

\item{plot}{logical, whether to plot GOF plots}

\item{return_values}{logical, whether to return GOF values}
}
\description{
Plot the compensator ans other associated goodness of fit metrics for a Hawkes process
}
