% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{show_multivariate_hawkes}
\alias{show_multivariate_hawkes}
\title{Multivariate Hawkes fitted model plot}
\usage{
show_multivariate_hawkes(obj, type = c("fitted", "data", "both"))
}
\arguments{
\item{obj}{Either object returned by \code{\link{fit_hawkes}}/\code{\link{fit_hawkes_cbf}} 
or a named list with elements \code{times} and \code{params}. If the latter, 
then \code{times} should be a numeric vector of observed time points, and 
\code{params} must contain, \code{alpha} (intensity jump after an event occurrence) 
and \code{beta} (exponential intensity decay). In addition, should contain 
either \code{mu} (base rate of the Hawkes process) or \code{background_parameters}
(parameter(s) for the assumed non-homogeneous background function;
could be a list of multiple values). May also contain \code{marks} (a vector of numerical marks).}

\item{type}{One of \code{c("fitted", "data", "both")}, default \code{"fitted"}.}
}
\description{
Multivariate Hawkes fitted model plot
}
