% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hawkes.r
\docType{methods}
\name{sim.hawkes}
\alias{sim.hawkes}
\title{Function to simulate a self-exciting Hawkes process.}
\usage{
sim.hawkes(mu, alpha, beta, n, plot = FALSE, seed = 123, method = "1")
}
\arguments{
\item{mu}{numeric integer specifying base rate of the hawkes process}

\item{alpha}{numeric integer specifying intensity jump after an event occurence}

\item{beta}{numeric integer specifying exponential intensity decay}

\item{n}{numeric depending on method 1) specifying end of the time line within which to simulate the proces, or 2)
specifying the number of timepoints to simulate.}

\item{plot}{logical if TRUE the process is plotted along with the intensity, by default is FALSE}

\item{seed}{seed by default = 123}

\item{method}{a character "1" or "2" specifying the method (see above) to simulate hawkes process}
}
\description{
There are two methods to simulate the process 1) sumiualtes a univariate hawkes process as
R  package 'hawkes' 2) as per algorithm defined https://radhakrishna.typepad.com/mle-of-hawkes-self-exciting-process.pdf
}
