% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lgcp.r
\name{sim_lgcp}
\alias{sim_lgcp}
\title{Function to simulate a spatiotemporal log-Gaussian Cox process using TMB and the
R_inla namespace for the spde construction of the latent field from a fitted model
or from a template}
\usage{
sim_lgcp(x, locs, temp.idx, mesh, parameters, covs, sp)
}
\arguments{
\item{x}{a fitted model from a call to \code{fit_lgcp}}
}
\description{
Function to simulate a spatiotemporal log-Gaussian Cox process using TMB and the
R_inla namespace for the spde construction of the latent field from a fitted model
or from a template
}
