% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_study.R
\name{sim_study}
\alias{sim_study}
\title{Sim study}
\usage{
sim_study(n, times, locs, param, domain, smesh, time_independent = F)
}
\arguments{
\item{n}{}

\item{times}{}

\item{locs}{}

\item{param}{}

\item{domain}{}

\item{smesh}{}

\item{time_independent}{}
}
\description{
Sim study
}
