% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{stelfi_load_inla}
\alias{stelfi_load_inla}
\title{Load INLA}
\usage{
stelfi_load_inla()
}
\value{
No return value. A message is printed in the console advising if \code{INLA}
has been successfully loaded.
}
\description{
Loads the INLA package with \code{requireNamespace("INLA", quietly = TRUE)}.
On loading \code{\link{stelfi}} \code{repos['INLA']} set as
\code{"https://inla.r-inla-download.org/R/testing/"}.
}
\references{
Lindgren, F. and Rue, H. (2015) Bayesian spatial modelling with R-INLA.
\emph{Journal of Statistical Software}, \strong{63}: 1--25.
}
