% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{terrorism}
\alias{terrorism}
\title{Terrorism in Iraq 2013--2017}
\format{
A data frame with 4208 rows and 16 variables:
\describe{
\item{iyear}{numeric year 2013--2017}
\item{imonth}{numeric month index 1--12}
\item{iday}{numeric day 1--31 (zeros are a non-entry)}
\item{country}{country (IRAQ)}
\item{latitude}{latitude location}
\item{longitude}{longitude location}
\item{sucess}{Logical fatal or not. TRUE = fatal}
\item{nkill}{number of fatalities per attack}
\item{specificity}{spatial accuracy of event: 1 = most accurate, 5 = worst}
\item{gname}{character name of attack perpetrators (ISIL)}
\item{x.coord}{x coordinate from location projected onto a sphere}
\item{y.coord}{y coordinate from location projected onto a sphere}
\item{z.coord}{z coordinate from location projected onto a sphere}
\item{popdensity}{scaled: number of people per kilometer squared}
\item{luminosity}{scaled: luminosity}
\item{tt}{scaled: time to nearest city in minutes}
}
}
\source{
\url{http://www.start.umd.edu/gtd/}
}
\description{
A dataset containing information of terrorism activity carried out
by the Islamic State of Iraq and the Levant (ISIL) in Iraq,
2013--2017
}
