% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshmetrics.r
\name{tri_area}
\alias{tri_area}
\title{Sin function to calculate the area of a triangle given
edge lengths a, b and angle C.}
\usage{
tri_area(a, b, angC)
}
\arguments{
\item{a}{A numeric triangle edgelength (opposite angle A).}

\item{b}{A numeric triangle edgelength (opposite angle B).}

\item{angC}{angle (in degrees)}
}
\description{
Sin function to calculate the area of a triangle given
edge lengths a, b and angle C.
}
\examples{
\dontrun{
A <- c(1, 1)
B <- c(2, 1)
C <- c(2.5, 1.5)
a <- dist(B, C)
b <- dist(A, C)
c <- dist(A, B)
angC <- ang( a, b, c)
tri_area(a, b, angC)
}
}
