% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{ufo}
\alias{ufo}
\title{UFO sightings in the USA}
\format{
A data frame with 97027 observations and 12 variables
\describe{
\item{city}{city where sighting was reported}
\item{state}{state code where sighting was reported}
\item{date_time}{\code{POSIXct}, time and date of sighting}
\item{city_latitude}{latitude of sighting}
\item{city_longitude}{longitude of sighting}
}
}
\source{
\url{https://data.world/timothyrenner/ufo-sightings#}
}
\description{
Subset of data sourced from the UFO Sightings Map
(\url{https://www.arcgis.com/apps/webappviewer/index.html?id=ddda71d5211f47e782b12f3f8d06246e}).
}
