% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{uk_serial}
\alias{uk_serial}
\title{Serial killers of the UK, 1828 - 2015}
\format{
A dataframe with 62 rows and 8 variables:
\describe{
\item{number_of_kills}{approx number of murders committed}
\item{years}{The years of operation}
\item{name}{Name of convicted serial killer}
\item{aka}{Some serial killers were given nicknames}
\item{year_start}{Year the murders began}
\item{tear_end}{Year the murders ended}
\item{date_of_first_kill}{Date, if known, first murder was committed}
\item{population_million}{Est popn in millions at time of first murder}
}
}
\source{
\url{https://www.murderuk.com/}
}
\description{
A dataset containing the names and number of recorded
murders committed by some of the infamous UK serial killers
from 1828 to 2015.
}
