% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{xyt}
\alias{xyt}
\title{A self-exciting spatiotemporal point pattern of class \code{stppp}}
\format{
A \code{stppp} object with 653 observations
\describe{
\item{window}{domain of the point pattern of class \code{owin}}
\item{n}{number of observations, 653}
\item{x}{x coordinate}
\item{y}{y coordinate}
\item{markformat}{\code{none}}
\item{t}{timestamp of points}
\item{tlim}{time frame \code{0 2}}
}
}
\description{
A self-exciting spatiotemporal point pattern of class \code{stppp}
}
