% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.r
\name{compile_stelfi}
\alias{compile_stelfi}
\title{Fuction to compile TMB C++ templates}
\usage{
compile_stelfi()
}
\description{
Compiles the  TMB templates into a shared object file.
Must be done a single time following installation or updating of the package.
}
