% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.r
\name{dll_stelfi}
\alias{dll_stelfi}
\title{Function to load DLLs for C++ templates
Loads required DLLs for models fitted using TMB}
\usage{
dll_stelfi(x)
}
\arguments{
\item{x}{option, if provided specifies \code{stelfi} DLL to load}
}
\description{
Function to load DLLs for C++ templates
Loads required DLLs for models fitted using TMB
}
