% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{earthquakes}
\alias{earthquakes}
\title{A dataset taken from the GeoNet Quake (http://quakesearch.geonet.org.nz/)
containing earthquake information in Canterbury, New Zealand 16-Jan-2010--24-Dec-2014}
\format{A data frame containing 3824 rows and 5 variables:
\describe{
\item{origintime}{The UTC time of the event's occurrence}
\item{longitude}{longitude location}
\item{latitude}{latitude location}
\item{magnitude}{The magnitude of the earthquake}
\item{depth}{The focal depth of the event (km)}
}}
\usage{
data(earthquakes)
}
\description{
A dataset taken from the GeoNet Quake (http://quakesearch.geonet.org.nz/)
containing earthquake information in Canterbury, New Zealand 16-Jan-2010--24-Dec-2014
}
