% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hawkes.r
\name{fit_hawkes}
\alias{fit_hawkes}
\title{Self-exciting Hawkes process}
\usage{
fit_hawkes(
  times,
  parameters = list(),
  model = 1,
  marks = c(rep(1, length(times))),
  tmb_silent = TRUE,
  optim_silent = TRUE,
  ...
)
}
\arguments{
\item{times}{A vector of numeric observed time points.}

\item{parameters}{A named list of named parameters for the chosen model:
\itemize{
\item \code{mu}, base rate of the Hawkes process,
\item \code{alpha} (supplied if \code{model = 1}), intensity jump after an event occurrence,
\item \code{a_par} (supplied if \code{model} = 2), logit scale for \code{alpha} (must be 
set so that the intensity never goes negative and so that \code{alpha} <= \code{beta}), and
\item \code{beta}, exponential intensity decay.
}}

\item{model}{A numeric indicator specifying which model to fit:
\itemize{
\item \code{1}, Hawkes process with exponential decay (default);
\item \code{2}, Hawkes process with an \code{alpha} that can be negative.
}}

\item{marks}{Optional, a vector of numeric marks, defaults to 1 (i.e., no marks).}

\item{tmb_silent}{Logical, default `TRUE`:
TMB inner optimization tracing information will be printed.}

\item{optim_silent}{Logical, default `TRUE`:
print function and parameters every iteration.}

\item{...}{arguments to pass into \code{optim()}}
}
\description{
`fit_hawkes()` fits a self-exciting Hawkes process using \code{TMB}.
}
\examples{
{
\dontrun{
## simple Hawkes model
data(retweets_niwa, package = "stelfi")
times <- unique(sort(as.numeric(difftime(retweets_niwa, min(retweets_niwa), units = "mins"))))
params <- c(mu = 9, alpha = 3, beta = 10)
fit_hawkes(times = times, parameters = params)
## A model with marks (ETAS-type)
data("earthquakes", package = "stelfi")
earthquakes <- earthquakes[order(earthquakes$origintime),]
earthquakes <- earthquakes[!duplicated(earthquakes$origintime), ]
times <- earthquakes$origintime
times <- as.numeric(difftime(times, min(times), units = "mins"))
marks <- earthquakes$magnitude
params <- c(mu = 3, alpha = 0.05, beta = 1)
fit_hawkes(times = times, parameters = params, marks = marks)
}
}
}
