% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hawkes.r
\name{fit_hawkes}
\alias{fit_hawkes}
\title{Function to fit a self-exciting Hawkes process}
\usage{
fit_hawkes(times, parameters, ...)
}
\arguments{
\item{times}{a vector of numeric observation time points}

\item{parameters}{a vector of named parmeters for the hawkes process: "mu"--base rate of the hawkes process,
"alpha"--intensity jump after an event occurence, and "beta"--exponential intensity decay}

\item{...}{arguments to pass into \code{nlminb()}}
}
\description{
Function to fit a self-exciting Hawkes process
}
