% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hawkes.r
\name{fit_hawkes_cbf}
\alias{fit_hawkes_cbf}
\title{Function to fit a self-exciting Hawkes process with a given custom
background function}
\usage{
fit_hawkes_cbf(
  times,
  parameters = list(),
  model = 1,
  marks = c(rep(1, length(times))),
  background,
  background_integral,
  background_parameters,
  background_min,
  tmb_silent = TRUE,
  optim_silent = TRUE
)
}
\arguments{
\item{times}{a vector of numeric observation time points}

\item{parameters}{a list of named parameters for the chosen model
Default values used if not provided. 
(see \code{model})
"mu"--base rate of the hawkes process,
"alpha"--intensity jump after an event occurrence (\code{model} = 1)
"a_par" -- logit scale for alpha. alpha must be set so that the intensity never goes negative
 and so that alpha <= beta. (\code{model} = 2)
"beta"--exponential intensity decay}

\item{model}{A factor indicator specifying which model to fit:
\itemize{
\item \code{1}, a Hawkes process with exponential decay and cbf (default),
positive alpha.
\item \code{2}, a Hawkes process with exponential decay and cbf,
but alpha can be negative.
}}

\item{marks}{a vector of numerical marks, defaults to 1 (i.e., no marks)}

\item{background}{A function taking one parameter and an
independent variable, returning a scalar.}

\item{background_integral}{The integral of \code{background}.}

\item{background_min}{A function taking one parameter and two points,
returns min of \code{background} between those points.}

\item{tmb_silent}{logical, default `TRUE`:
TMB inner optimization tracing information will be printed.}

\item{optim_silent}{logical, default `TRUE`:
print function and parameters every iteration.}

\item{background_parameter}{The parameter(s)
for the background function \code{background}.
This could be a list of multiple values.}
}
\description{
\code{fit_hawkes_cbf} fits a range of self-exciting Hawkes processes
with a given custom background function (cbf) using TMB using TMB.
The \code{alpha} (\code{model} = 1) or \code{a_par} (\code{model} = 2)
and \code{beta} parameters are estimated using TMB,
parameters of the cbf are optimized in R.
}
\examples{
{
\dontrun{
times <- hawkesbow::hawkes(1000, fun = function(y) {1 + 0.5*sin(y)},
M = 1.5, repr = 0.5, family = "exp", rate = 2)$p
## The background function must take a single parameter and
## the time(s) at which it is evaluated
background <- function(params,times) {
A = exp(params[[1]])
B = stats::plogis(params[[2]]) * A
return(A + B  *sin(times))
}
## The background_integral function must take a
## single parameter and the time at which it is evaluated
background_integral <- function(params,x) {
        A = exp(params[[1]])
        B = stats::plogis(params[[2]]) * A
        return((A*x)-B*cos(x))
}
param = list(alpha = 0.5, beta = 1.5)
background_param = list(1,1)
fit_hawkes_cbf(times = times, parameters = param,
background = background,
background_integral = background_integral,
background_parameters = background_param)
}
}
}
