% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stelfi.r
\name{fit_hspat_tmb}
\alias{fit_hspat_tmb}
\title{Function to fit a spatial Hawkes process using \code{TMB}}
\usage{
fit_hspat_tmb(
  times,
  locs,
  sf,
  smesh,
  coefs,
  designmat,
  logit_abratio = 0,
  log_beta = 0,
  log_xsigma = 0,
  log_ysigma = 0,
  atanh_rho = 0,
  w,
  reltol = 1e-12,
  abstol = 1e-12,
  tmax = max(times),
  lmat,
  simple,
  tmb_silent,
  nlminb_silent,
  ...
)
}
\arguments{
\item{times}{a vector of numeric observation time points}

\item{locs}{A \code{data.frame} of \code{x} and \code{y} locations, 2xn. If locations have
time stamps then this should be the third column of the supplied 3xn matrix.}

\item{sf}{An \code{sf} of type \code{POLYGON} specifying the region
of the domain.}

\item{smesh}{A spatial mesh of class \code{\link[INLA]{inla.mesh.2d}}.}

\item{coefs}{TODO}

\item{designmat}{TODO}

\item{logit_abratio}{logit ration of a, b parameters of the temporal Hawkes process}

\item{log_beta}{the \code{log(beta)} base rate of the Hawkes process}

\item{log_xsigma}{TODO}

\item{log_ysigma}{TODO}

\item{atanh_rho}{optional, \code{arctan(rho)} AR1 parameter.}

\item{w}{A vector of model weights; corresponds to the \code{E} term for
poisson models, see \code{INLA::inla.doc("poisson")} for more detail.}

\item{reltol}{\code{numeric}, relative tolerance (default  \code{1e-12})}

\item{abstol}{\code{numeric}, absolute tolerance (default  \code{1e-12})}

\item{tmax}{TODO}

\item{lmat}{TODO}

\item{simple}{TODO}

\item{tmb_silent}{logical, default `TRUE`:
TMB inner optimization tracing information will be printed.}

\item{nlminb_silent}{Logical, default \code{TRUE}:
print function and parameters every iteration.}

\item{...}{arguments to pass into \code{optim()}}
}
\description{
Function to fit a spatial Hawkes process using \code{TMB}
}
