% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lgcp.r
\name{fit_lgcp}
\alias{fit_lgcp}
\title{Function to fit a spatial or spatiotemporal log-Gaussian Cox process using \code{TMB}}
\usage{
fit_lgcp(locs, sp, smesh, tmesh, parameters, covariates, silent, ...)
}
\arguments{
\item{locs}{2xn \code{data.frame} of locations x, y. If locations have
time stapms then this is the third column of the 3xn matrix}

\item{sp}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} of the domain}

\item{smesh}{spatial mesh of class \code{"inla.mesh"}}

\item{tmesh}{optional, temporal mesh of class \code{"inla.mesh.1d"}}

\item{parameters}{a list of named parmeters:
"beta"--A vector of fixed effects coefficients to be estimated
(same length as \code{ncol(covariates)} + 1 )
 "tau"-- tau parameter for the GMRF
 "kappa"-- kappa parameter for the GMRF
 "rho"-- optional, rho AR1 parameter}

\item{covariates}{optional, a \code{matrix} of covariates at each
\code{smesh} and \code{tmesh} node combination.}

\item{silent}{logical, by default FALSE. If TRUE model fitting progress
not printed to console.}

\item{...}{arguments to pass into \code{nlminb()}}
}
\description{
A simple to use wrapper for \code{fit_lgcp_tmb}
}
