% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mlgcp.r
\name{fit_mlgcp}
\alias{fit_mlgcp}
\title{Function to fit a spatial  marked
log-Gaussian Cox process using \code{TMB}}
\usage{
fit_mlgcp(
  locs,
  sp,
  marks,
  smesh,
  parameters,
  methods,
  strfixed,
  strparam,
  idx,
  ...
)
}
\arguments{
\item{locs}{2xn \code{data.frame} of locations x, y. If locations have
time stapms then this is the third column of the 3xn matrix}

\item{sp}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} of the domain}

\item{smesh}{spatial mesh of class \code{"inla.mesh"}}

\item{parameters}{a list of named parmeters:
"beta"--A vector of  coefficients to be estimated
of length \code{ncol(marks)} + 1.
 "tau"-- tau parameter for the GMRF.
 "kappa"-- kappa parameter for the GMRF.
\code{smesh} and \code{tmesh} node combination.}

\item{idx}{a binary vector of the same size as the observation
vector \code{y}.
With this vector, the log-likelihood can be computed using a subset
of the observations. 1 for contributing to the log-likelihood,
and 0 otherwise.}

\item{...}{arguments to pass into \code{nlminb()}}
}
\description{
A simple to use wrapper for \code{fit_mlgcp_tmb}
}
