% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mlgcp.r
\name{fit_mlgcp}
\alias{fit_mlgcp}
\title{Marked spatial log-Gaussian Cox process m(LGCP)}
\usage{
fit_mlgcp(
  locs,
  sf,
  marks,
  smesh,
  parameters = list(),
  methods,
  strfixed = matrix(1, nrow = nrow(locs), ncol = ncol(marks)),
  fields = rep(1, ncol(marks)),
  covariates,
  pp_covariates,
  marks_covariates,
  tmb_silent = TRUE,
  nlminb_silent = TRUE,
  ...
)
}
\arguments{
\item{locs}{A \code{data.frame} of \code{x} and \code{y} locations, 2xn.}

\item{sf}{An \code{sf} of type \code{POLYGON} specifying the region
of the domain.}

\item{marks}{A matrix of marks for each observation of the point pattern}

\item{smesh}{A spatial mesh of class \code{\link[INLA]{inla.mesh.2d}}.}

\item{parameters}{a list of named parameters:
log_tau, log_kappa, betamarks, betapp, marks_coefs_pp.
See \code{\link{fit_mlgcp_tmb}} for further details.}

\item{methods}{An integer value:
\itemize{
\item \code{0} (default), Gaussian distribution, parameter estimated is mean;
\item \code{1}, Poisson distribution, parameter estimated is intensity;
\item \code{2}, binomial distribution, parameter estimated is logit/probability;
\item \code{3}, gamma distribution, the implementation in TMB is shape-scale.
}}

\item{strfixed}{A matrix of fixed structural parameters, defined for each event and mark.
Defaults to 1s.
Normal distribution: this is the log of standard deviation.
Poisson distribution: not used
Binomial distribution, this is the number of trials.
Gamma distribution, this is the log of the scale.}

\item{fields}{A binary vector indicating whether there is a new random
field for each mark. By default, each mark has its own random field.}

\item{covariates}{Covariate(s) corresponding to each area in the spatial mesh}

\item{pp_covariates}{Which columns of the covariates apply to the point process}

\item{marks_covariates}{Which columns of the covariates apply to the marks.
By default, all covariates apply to the marks only.}

\item{tmb_silent}{Logical, default \code{TRUE}:
TMB inner optimization tracing information will be printed.}

\item{nlminb_silent}{Logical, default \code{TRUE}:
print function and parameters every iteration.}

\item{...}{optional extra arguments to pass into \code{\link[stats]{nlminb}}.}
}
\description{
`fit_mlgcp_tmb()` fits a marked LGCP using \code{TMB} and the
\code{R_inla} namespace for the spde construction of the latent field.
A simple to use wrapper for \code{\link{fit_mlgcp_tmb}}.
}
\examples{
{
\dontrun{
data(marked, package = "stelfi")
loc.d <- 3 * cbind(c(0, 1, 1, 0, 0), c(0, 0, 1, 1, 0))
domain <- sf::st_sf(geometry = sf::st_sfc(sf::st_polygon(list(loc.d))))
smesh <- INLA::inla.mesh.2d(loc.domain = loc.d, offset = c(0.3, 1),
max.edge = c(0.3, 0.7), cutoff = 0.05)
locs <- cbind(x = marked$x, y = marked$y)
marks <- cbind(m1 = marked$m1) ## Gaussian
parameters <- list(betamarks = matrix(0, nrow = 1, ncol = ncol(marks)),
log_tau = rep(log(1), 2), log_kappa = rep(log(1), 2),
marks_coefs_pp = rep(0, ncol(marks)), betapp = 0)
fit <- fit_mlgcp(locs = locs, marks = marks,
sf = domain, smesh = smesh,
parameters = parameters, methods = 0,fields = 1)
}
}
}
