% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mlgcp.r
\name{fit_mlgcp}
\alias{fit_mlgcp}
\title{Function to fit a spatial  marked
log-Gaussian Cox process using \code{TMB}}
\usage{
fit_mlgcp(
  locs,
  sp,
  marks,
  smesh,
  parameters,
  methods,
  strfixed,
  fields,
  covariates,
  pp_covariates,
  marks_covariates,
  tmb_silent = TRUE,
  nlminb_silent = TRUE,
  ...
)
}
\arguments{
\item{locs}{2xn \code{data.frame} of locations x, y. If locations have
time stapms then this is the third column of the 3xn matrix}

\item{sp}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} of the domain}

\item{smesh}{spatial mesh of class \code{"inla.mesh"}}

\item{parameters}{a list of named parmeters:
log_tau, log_kappa, betamarks, betapp, marks_coefs_pp. 
See fit_mlgcp_tmb header for further details
\code{smesh} and \code{tmesh} node combination.}

\item{fields}{See fit_mlgcp_tmb header}

\item{covariates}{Covariate(s) corresponding to each area in the spatial mesh}

\item{pp_covariates}{Which columns of the covariates apply to the point process}

\item{marks_covariates}{Which columns of the covariates apply to the marks.
By default, all covariates apply to the marks only.}

\item{tmb_silent}{Logical, default \code{TRUE}:
TMB inner optimization tracing information will be printed.}

\item{nlminb_silent}{logical, default \code{TRUE}:
print function and parameters every iteration.}

\item{...}{Optional extra arguments to pass into \code{\link[stats]{nlminb}}.}
}
\description{
A simple to use wrapper for \code{fit_mlgcp_tmb}
}
