% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mlgcp.r
\name{fit_mlgcp_tmb}
\alias{fit_mlgcp_tmb}
\title{Function to fit a marked spatial log-Gaussian Cox
process (LGCP) using \code{TMB} and the
\code{R_inla} namespace for the spde construction of the latent field.
For a simpler wrapper use \code{fit_mlgcp()}}
\usage{
fit_mlgcp_tmb(
  ypp,
  marks,
  lmat,
  spde,
  w,
  strfixed,
  methods,
  betamarks,
  betapp,
  marks_coefs_pp,
  log_kappa,
  log_tau,
  cov_overlap,
  designmatpp,
  designmatmarks,
  fields,
  tmb_silent,
  nlminb_silent,
  ...
)
}
\arguments{
\item{ypp}{A vector for point process response (poisson) (see \code{fit_lgcp()})}

\item{marks}{A matrix of marks for each observation of the point pattern}

\item{lmat}{A sparse matrix mapping mesh points to the observations}

\item{spde}{The structure of spde object as defined in \code{r-inla}.
The minimal required components are M0, M1, M2.}

\item{w}{a vector of model weights; corresponds to the E term for
poisson models in \code{INLA},
see \code{INLA::inla.doc("poisson")} for more detail.}

\item{strfixed}{a matrix of fixed structural parameters.
Normal distribution: this is the log of element-wise standard deviation.
Poisson distribution: not used
Binomial distribution, this is the number of trials.
The default input from R should be 1s, i.e., the Bernoulli distribution.
Gamma distribution, this is the log of scale.}

\item{methods}{integer:
0 - normal distribution. Parameter estimated is mean
1 - poisson distribution. Parameter estimated is intensity
2 - binomial distribution. Parameter estimated is logit/probability
3 - gamma distribution, the implementation in TMB is shape-scale.}

\item{betamarks}{starting value of the intercept term and covariate coefficient of each mark}

\item{betapp}{starting value of the  intercept of point process}

\item{marks_coefs_pp}{slope of mark lambda versus point process lambda}

\item{log_kappa}{log of kappas for the random field(s). The first element is for the random field of the point process.}

\item{log_tau}{log of taus for the random field(s). The first element is for the random field of the point process.}

\item{cov_overlap}{whether or nor marks and point process share covariates. If not, marks_coef_pp applies to entire lambda of the point process.
If there is overlap, it applies only to the point process random field.}

\item{designmatpp}{Design matrix for point process. The first column is ones.
If there are covariates, then the covariates are in the subsequent columns}

\item{designmatmarks}{Design matrix for marks}

\item{fields}{Binary vector indicating whether there is a new random field for each mark.}

\item{tmb_silent}{logical, default `TRUE`:
TMB inner optimization tracing information will be printed.}

\item{nlminb_silent}{logical, default `TRUE`:
print function and parameters every iteration.}

\item{...}{arguments to pass into \code{nlminb()}}
}
\description{
Function to fit a marked spatial log-Gaussian Cox
process (LGCP) using \code{TMB} and the
\code{R_inla} namespace for the spde construction of the latent field.
For a simpler wrapper use \code{fit_mlgcp()}
}
