% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stelfi.r
\name{fit_stelfi}
\alias{fit_stelfi}
\title{Funtion to fit stelfi}
\usage{
fit_stelfi(
  times,
  locs,
  sp,
  smesh,
  parameters,
  gaussian = TRUE,
  tmb_silent = TRUE,
  nlminb_silent = TRUE,
  ...
)
}
\arguments{
\item{times}{a vector of numeric observation time points}

\item{locs}{2xn \code{data.frame} of locations x, y. If locations have
time stapms then this is the third column of the 3xn matrix}

\item{sp}{\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} of the domain}

\item{tmb_silent}{logical, default `TRUE`:
TMB inner optimization tracing information will be printed.}

\item{nlminb_silent}{logical, default `TRUE`:
print function and parameters every iteration.}

\item{...}{arguments to pass into \code{optim()}}
}
\description{
Funtion to fit stelfi
}
