% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_coefs}
\alias{get_coefs}
\title{EXtract reported parameter estimates from fitted models}
\usage{
get_coefs(object)
}
\arguments{
\item{object}{result of a call to \code{\link{fit_hawkes}} or \code{\link{fit_lgcp}}}
}
\description{
\code{get_coefs} returns the parameter estimates for the fitted model.
}
