% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_fields}
\alias{get_fields}
\title{Estimated random field(s)}
\usage{
get_fields(object, smesh, tmesh, plot = FALSE, sd = FALSE)
}
\arguments{
\item{object}{The result of a call to \code{\link{fit_lgcp}}.}

\item{plot}{Logical, if \code{TRUE} then the returned values are plotted.}

\item{sd}{Logical, if \code{TRUE} then standard errors of field are returned.}
}
\description{
\code{get_fields} returns
}
