% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_fields}
\alias{get_fields}
\title{Estimated random field(s)}
\usage{
get_fields(object, smesh, tmesh, plot = FALSE, sd = FALSE)
}
\arguments{
\item{object}{result of a call \code{fit_lgcp()}}

\item{smesh}{spatial mesh of class \code{"inla.mesh"}}

\item{tmesh}{optional, temporal mesh of class \code{"inla.mesh.1d"}}

\item{plot}{logical, if TRUE then field(s) plotted}

\item{sd}{logical, if TRUE then standard errors of field returned}
}
\description{
Estimated random field(s)
}
