% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_fields}
\alias{get_fields}
\title{Function to extract fields from a fitted model (INLA only ATM) (spatial only ATM)}
\usage{
get_fields(x = "inla", mesh = "inla.mesh")
}
\description{
Function to extract fields from a fitted model (INLA only ATM) (spatial only ATM)
}
