% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_fields}
\alias{get_fields}
\title{Function to extract fields from a fitted model (INLA only ATM)}
\usage{
get_fields(
  x = "inla",
  mesh = "inla.mesh",
  t = "numeric_or_NULL",
  mean = "logical"
)
}
\arguments{
\item{x}{an \code{inla} object}

\item{mesh}{an object of class \code{inla.mesh}}

\item{t}{optional, if supplied specifies the number of time steps}

\item{mean}{logical, if TRUE extracts point estimates of the fields,
else returns the standard errors}
}
\description{
Function to extract fields from a fitted model (INLA only ATM)
}
