% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_weights}
\alias{get_weights}
\title{Function to find areas (weights) around the mesh nodes which are
within the specified spatial polygon.
Relies on the internal \code{\link{inla.mesh.dual}} function.}
\usage{
get_weights(mesh, sp, plot = FALSE)
}
\description{
Function to find areas (weights) around the mesh nodes which are
within the specified spatial polygon.
Relies on the internal \code{\link{inla.mesh.dual}} function.
}
\seealso{
\url{https://becarioprecario.bitbucket.io/spde-gitbook/}.
}
