% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_weights}
\alias{get_weights}
\title{Function to find areas (weights) around the mesh nodes which are
within the specified spatial polygon.
Relies on inla.mesh.dual function from INLA spde-tutorial}
\usage{
get_weights(mesh = NULL, boundary = NULL)
}
\description{
Function to find areas (weights) around the mesh nodes which are
within the specified spatial polygon.
Relies on inla.mesh.dual function from INLA spde-tutorial
}
