% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{hawke_intensity}
\alias{hawke_intensity}
\title{Hawkes intensty function with decay historical dependence}
\usage{
hawke_intensity(mu, alpha, beta, times, p = NULL)
}
\arguments{
\item{times}{a vector of numeric observation time points}
}
\description{
Hawkes intensty function with decay historical dependence
}
