% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{hawke_intensity}
\alias{hawke_intensity}
\title{Hawkes intensity function}
\usage{
hawke_intensity(times, mu, alpha, beta, p, marks, background_param)
}
\arguments{
\item{times}{A numeric vector of observed time points.}

\item{mu}{numeric integer specifying base rate of the hawkes process}

\item{alpha}{numeric integer specifying intensity jump after an event
occurence}

\item{beta}{numeric integer specifying exponential intensity decay}

\item{p}{An optional vector of pseudo times at which to calculate the intensity.}
}
\description{
Hawkes intensity function
}
