% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{hawke_intensity}
\alias{hawke_intensity}
\title{Hawkes intensty function with decay historical dependence}
\usage{
hawke_intensity(mu, alpha, beta, times, p)
}
\arguments{
\item{mu}{numeric integer specifying base rate of the hawkes process}

\item{alpha}{numeric integer specifying intensity jump after an event occurence}

\item{beta}{numeric integer specifying exponential intensity decay}

\item{times}{a vector of numeric observation time points}

\item{p}{vector of pseudo times at which to calculate the intensity}
}
\description{
Hawkes intensty function with decay historical dependence
}
