% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_wrappers.r
\name{inla.mesh.dual}
\alias{inla.mesh.dual}
\title{Function copied from \url{http://www.r-inla.org/spde-book}
which is suppied alongside the
\href{SPDE gitbook}{https://becarioprecario.bitbucket.io/spde-gitbook/}}
\source{
\url{http://www.r-inla.org/spde-book}
}
\usage{
inla.mesh.dual(mesh)
}
\description{
Function copied from \url{http://www.r-inla.org/spde-book}
which is suppied alongside the
\href{SPDE gitbook}{https://becarioprecario.bitbucket.io/spde-gitbook/}
}
