% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hawkes.r
\name{inla.mesh.dual}
\alias{inla.mesh.dual}
\title{Function copied from http://inla.r-inla-download.org/r-inla.org/tutorials/spde/R/spde-tutorial-functions.R}
\usage{
inla.mesh.dual(mesh)
}
\description{
Function copied from http://inla.r-inla-download.org/r-inla.org/tutorials/spde/R/spde-tutorial-functions.R
}
