% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshmetrics.r
\name{mesh_ang}
\alias{mesh_ang}
\title{Calculate all interior Delaunay triangulation triangle angles}
\usage{
mesh_ang(mesh)
}
\arguments{
\item{mesh}{A \code{\link[INLA]{inla.mesh.2d}} object.}

\item{s}{A dataframe of Delaunay triangulation segments as returned by
\code{\link{segments}}.}
}
\description{
Calculate all interior Delaunay triangulation triangle angles
}
\examples{
\dontrun{
data(horse_mesh, package = "stelfi")
s <- segments(horse_mesh)
mesh_ang(horse_mesh, s)
}
}
