% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshmetrics.r
\name{meshmetrics}
\alias{meshmetrics}
\title{Calculate a number of different \code{\link[INLA]{inla.mesh.2d}} attributes}
\usage{
meshmetrics(mesh)
}
\arguments{
\item{mesh}{A \code{\link[INLA]{inla.mesh.2d}} object.}
}
\value{
A named list:
\itemize{ 
\item \code{edges}, a data frame of \code{mesh} segments as returned by \code{\link{segments}};
\item \code{triangles}, a data frame specifying, for each triangle in the triangulation, the
 \itemize{
\item circumcircle radius \code{circumcircle_R},
\item incircle radius \code{incircle_r}), 
\item the assocoated centroid locations (\code{c_Ox, cOy} and \code{i_Ox, iOy}), 
\item the radius-edge ratio \code{radius_edge}, and 
\item the radius ratio \code{radius_ratio};
}
\item \code{angles}, a 3 x nvert dataframe of all triangle interior angles.
}
}
\description{
\code{meshmetrics} calculates a number of gemoetric attributes for a given
Delaunay triangulation.
}
\examples{
\dontrun{
data(horse_mesh, package = "stelfi")
meshmetrics(horse_mesh)
}
}
