% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshmetrics.r
\name{meshmetrics}
\alias{meshmetrics}
\title{Calculate a number of different \code{\link[INLA]{inla.mesh.2d}} attributes}
\usage{
meshmetrics(mesh)
}
\arguments{
\item{mesh}{A \code{\link[INLA]{inla.mesh.2d}} object.}
}
\value{
An object of class \code{sf} with the following data:
\itemize{ 
\item \code{V1}, \code{V2}, and \code{V3} corresponding vertecies of \code{mesh}
matches \code{mesh$graph$tv}.
\item \code{ID} numeric triangle id
\item {angleA}, \code{angleB}, and \code{angleC} the interior angles of the triangles
\item circumcircle radius \code{circumcircle_R},
\item incircle radius \code{incircle_r}), 
\item the assocoated centroid locations (\code{c_Ox, cOy} and \code{i_Ox, iOy}), 
\item the radius-edge ratio \code{radius_edge}, and 
\item the radius ratio \code{radius_ratio}
\item \codea{area} triangle area
\item \code{quality} a measure of triangle "quality" defined as
\frac{4\sqrt{3}|A|}{\Sigma_{i = 1}^3 L_i^2},
where A is the area and L_i is edge length.
}
}
\description{
\code{meshmetrics} calculates a number of gemoetric attributes for a given
Delaunay triangulation.
}
\examples{
\dontrun{
data(horse_mesh, package = "stelfi")
meshmetrics(horse_mesh)
}
}
