% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_lambda}
\alias{plot_lambda}
\title{Plot the lambda estimate from a fitted LGCP model}
\usage{
plot_lambda(
  fit,
  covariates,
  smesh,
  tmesh,
  dims = c(500, 500),
  sf,
  timestamp = 1
)
}
\arguments{
\item{fit}{a fitted model object for, for example, \code{fit_lgcp()}.}

\item{covariates}{Optional, a \code{matrix} of covariates at each
\code{smesh} and \code{tmesh} node combination.}

\item{smesh}{A spatial mesh of class \code{\link[INLA]{inla.mesh.2d}}.}

\item{tmesh}{Optional, a temporal mesh of class \code{\link[INLA]{inla.mesh.1d}}.}

\item{dims}{2D spatial resolution to plot, default \code{c(500,500)}.}

\item{sf}{An \code{sf} of type \code{POLYGON} specifying the region
of the domain.}

\item{timestamp}{index of time stamp to plot, default 1.}
}
\description{
Plot the lambda estimate from a fitted LGCP model
}
\examples{
\dontrun{
data(xyt, package = "stelfi")
domain <- sf::st_as_sf(xyt$window)
locs <- data.frame(x = xyt$x, y = xyt$y)
bnd <- INLA::inla.mesh.segment(as.matrix(sf::st_coordinates(domain)[, 1:2]))
smesh <- INLA::inla.mesh.2d(boundary = bnd, 
max.edge = 0.75, cutoff = 0.3)
fit <- fit_lgcp(locs = locs, sf = domain, smesh = smesh,
parameters = c(beta = 0, log_tau = log(1), log_kappa = log(1)))
plot_lambda(fit, smesh = smesh, sf = domain)
}
}
\seealso{
\code{\link{show_field}} and \code{\link{get_fields}}
}
