% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_lambda}
\alias{plot_lambda}
\title{Plot the lambda estimate from a spatial model}
\usage{
plot_lambda(
  fit,
  covariates,
  smesh,
  tmesh,
  dims = c(500, 500),
  border,
  colour_option = "D",
  title = ""
)
}
\description{
Plot the lambda estimate from a spatial model
}
