% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{points_in_mesh}
\alias{points_in_mesh}
\title{Internal function that takes in a list of points
(and optionally a weight or covariate for each point)
and a mesh of polygons.}
\usage{
points_in_mesh(xy, dmesh, weights)
}
\arguments{
\item{xy}{A data frame of locations.}

\item{dmesh}{An object returned by \code{\link{dual_mesh}}.}

\item{weights}{A vector of weights/covariates for each point.}
}
\description{
Internal function that takes in a list of points
(and optionally a weight or covariate for each point)
and a mesh of polygons.
}
