% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lgcp.r
\name{prep_data_lgcp}
\alias{prep_data_lgcp}
\title{Internal function to prep data as per \code{link{INLA}} stack}
\usage{
prep_data_lgcp(locs, sp, smesh, tmesh)
}
\arguments{
\item{locs}{A \code{data.frame} of \code{x} and \code{y} locations, 2xn. If locations have
time stamps then this should be the third column of the supplied 3xn matrix.}

\item{sp}{A \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
of the domain.}

\item{smesh}{A spatial mesh of class \code{\link[INLA]{inla.mesh.2d}}.}

\item{tmesh}{Optional, a temporal mesh of class \code{\link[INLA]{inla.mesh.1d}}.}
}
\description{
Internal function to prep data as per \code{link{INLA}} stack
}
