% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{serial_uk}
\alias{serial_uk}
\title{Serial killers of the UK, 1828 - 2015}
\format{
A dataframe with 62 rows and 8 variables:
\describe{
\item{Number.of.kills}{approx number of murders committed}
\item{Years}{The years of operation}
\item{Name}{Name of convicted serial killer}
\item{AKA}{Some serial killers were given nicknames}
\item{Year.Start}{Year the murders began}
\item{Year.End}{Year the murders ended}
\item{Date.of.first.kill}{Date, if known, first murder was committed}
\item{population..M.}{Est popn in millions at time of first murder}
}
}
\source{
\url{http://www.murderuk.com/}
}
\description{
A dataset containing the names and number of recorded
murders commited by some infamous UK serial killers 1828 - 2015.
}
