% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{show_field}
\alias{show_field}
\title{Plot estimated random field(s) of a fitted LGCP}
\usage{
show_field(x, smesh, dims = c(500, 500), border)
}
\arguments{
\item{x}{A vector of values for each \code{smesh} node.}

\item{dims}{A numeric vector of length 2 specifying
the spatial pixel resolution. By default \code{= c(500,500)}.}

\item{border}{Optional, a \code{SpatialPolygons} object of the domain border.}
}
\description{
Plot estimated random field(s) of a fitted LGCP
}
\examples{
\dontrun{
data(xyt, package = "stelfi")
domain <- as(xyt$window, "SpatialPolygons")
smesh <- INLA::inla.mesh.2d(boundary = INLA::inla.sp2segment(domain), 
max.edge = 0.75, cutoff = 0.3)
parameters <- c(beta = 1, log_tau = log(1), log_kappa = log(1))
simdata <- simulate_lgcp(parameters = parameters, sp = domain, smesh = smesh)
show_field(simdata$x, smesh = smesh)
}
}
\seealso{
\code{\link{plot_lambda}} and \code{\link{get_fields}}
}
