% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\docType{methods}
\name{show_hawkes}
\alias{show_hawkes}
\title{Plots the Hawkes intensty function with decay historical dependence}
\usage{
show_hawkes(times, mu, alpha, beta)
}
\arguments{
\item{times}{a vector of numeric observation time points}
}
\description{
Plots the Hawkes intensty function with decay historical dependence
}
