% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{show_hawkes}
\alias{show_hawkes}
\title{Plot Hawkes intensity}
\usage{
show_hawkes(obj)
}
\arguments{
\item{obj}{Either object returned by \code{fit_hawkes()}/\code{fit_hawkes_cbf()} 
or a named list with elements \code{times} and \code{params}. If the latter, 
then \code{times} should be a numeric vector of observed time points, and 
\code{params} must contain, \code{alpha} (intensity jump after an event occurrence) 
and \code{beta} (exponential intensity decay). In addition, should contain 
either \code{mu} (base rate of the Hawkes process) or \code{background_parameters}
(parameter(s) for the assumed non-homogeneous background function;
could be a list of multiple values). 
May also contain \code{marks} (a vector of numerical marks).}
}
\description{
`show_hawkes()` plots a Hawkes intensity function, options to extend to 
non-homogeneous background intensity.
}
\examples{
\dontrun{
data(retweets_niwa, package = "stelfi")
times <- unique(sort(as.numeric(difftime(retweets_niwa, min(retweets_niwa),units = "mins"))))
params <- c(mu = 9, alpha = 3, beta = 10)
show_hawkes(list(times = times, params = params))
## must have compiled TMB templates first use compile_stelfi()
fit <- fit_hawkes(times = times, parameters = params)
show_hawkes(fit)
}
}
\seealso{
[hawkes_intensity()] [fit_hawkes_cbf()]
}
