% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{show_hawkes}
\alias{show_hawkes}
\title{Plots the Hawkes intensity function with decay historical dependence}
\usage{
show_hawkes(times, mu, alpha, beta, marks, obj, background_parameters)
}
\arguments{
\item{times}{A numeric vector of observed time points.}

\item{mu}{numeric integer specifying base rate of the hawkes process}

\item{alpha}{numeric integer specifying intensity jump after an event
occurence}

\item{beta}{numeric integer specifying exponential intensity decay}

\item{marks}{a vector of numerical marks, defaults to 1 (i.e., no marks)}

\item{obj}{An object returned by fit_hawkes() or fit_hawkes_cbf()
If provided, the only other parameters required are \code{mu} and \code{background_parameters}, and only if \code{mu} is not constant.}

\item{background_parameters}{The parameter(s)
for the background function \code{background}.
This could be a list of multiple values.}
}
\description{
Plots the Hawkes intensity function with decay historical dependence
}
\examples{
\dontrun{
data(retweets_niwa, package = "stelfi")
times <- unique(sort(as.numeric(difftime(retweets_niwa ,min(retweets_niwa),units = "mins"))))
params <- c(mu = 9, alpha = 3, beta = 10)
## must have compiled TMB templates first use compile_stelfi()
fit <- fit_hawkes(times = times, parameters = params)
pars <- get_coefs(fit)
show_hawkes(times = times, mu = pars[1,1], alpha = pars[2,1], beta = pars[3,1])
}
}
