% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{show_hawkes}
\alias{show_hawkes}
\title{Plots the Hawkes intensty function with decay historical dependence}
\usage{
show_hawkes(
  times,
  mu,
  alpha,
  beta,
  marks = c(rep(1, length(times))),
  background_param = NULL
)
}
\arguments{
\item{times}{A numeric vector of observed time points.}

\item{mu}{numeric integer specifying base rate of the hawkes process}

\item{alpha}{numeric integer specifying intensity jump after an event
occurence}

\item{beta}{numeric integer specifying exponential intensity decay}
}
\description{
Plots the Hawkes intensty function with decay historical dependence
}
\examples{
\dontrun{
data(retweets_niwa, package = "stelfi")
times <- unique(sort(as.numeric(difftime(retweets_niwa ,min(retweets_niwa),units = "mins"))))
params <- c(mu = 9, alpha = 3, beta = 10)
## must have compiled TMB templates first use compile_stelfi()
fit <- fit_hawkes(times = times, parameters = params) 
pars <- get_coefs(fit)
show_hawkes(times = times, mu = pars[1,1], alpha = pars[2,1], beta = pars[3,1])
}
}
