% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{show_hawkes_GOF}
\alias{show_hawkes_GOF}
\title{Plot the compensator and other associated goodness of fit metrics for a Hawkes process}
\usage{
show_hawkes_GOF(
  times,
  mu,
  alpha,
  beta,
  marks = c(rep(1, length(times))),
  background_param,
  plot = TRUE,
  return_values = TRUE
)
}
\arguments{
\item{times}{A numeric vector of observed time points.}

\item{mu}{numeric integer specifying base rate of the hawkes process}

\item{alpha}{numeric integer specifying intensity jump after an event
occurence}

\item{beta}{numeric integer specifying exponential intensity decay}

\item{plot}{Logical, whether to plot GOF plots. Default \code{TRUE}.}

\item{return_values}{Logical, whether to return GOF values. Default \code{TRUE}.}
}
\description{
Plot the compensator and other associated goodness of fit metrics for a Hawkes process
}
\examples{
\dontrun{
data(retweets_niwa, package = "stelfi")
times <- unique(sort(as.numeric(difftime(retweets_niwa ,min(retweets_niwa),units = "mins"))))
params <- c(mu = 9, alpha = 3, beta = 10)
## must have compiled TMB templates first use compile_stelfi()
fit <- fit_hawkes(times = times, parameters = params)
pars <- get_coefs(fit)
show_hawkes_GOF(times = times, mu = pars[1,1], alpha = pars[2,1], beta = pars[3,1], return_values = FALSE)
}
}
