% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{terrorism}
\alias{terrorism}
\title{A dataset taken from the global terrorism database (GTD) (http://www.start.umd.edu/gtd/)
containing information of terrorism activity 2010--2017}
\format{A data frame with 72366 rows and 16 variables:
\describe{
\item{iyear}{numeric year 2010--2017}
\item{imonth}{numeric month index 1--12}
\item{iday}{numeric day 1--31 (zeros are a non-entry)}
\item{country}{country}
\item{latitude}{latitude location}
\item{longitude}{longitude location}
\item{sucess}{Logical fatal or not. TRUE = fatal}
\item{nkill}{number of fatalities per attack}
\item{specificity}{factor which represents the apatial accuracy of the evets: 1 = most accurate, 5 = worst}
\item{gname}{character name of attack perpetrators}
\item{x.coord}{x coordinate from location projected onto a sphere}
\item{y.coord}{y coordinate from location projected onto a sphere}
\item{z.coord}{z coordinate from location projected onto a sphere}
\item{popdensity}{scaled: number of people per kilometer squared}
\item{luminosity}{scaled: luminosity}
\item{tt}{scaled: time to nearest city in minutes}
}}
\usage{
data(terrorism)
}
\description{
A dataset taken from the global terrorism database (GTD) (http://www.start.umd.edu/gtd/)
containing information of terrorism activity 2010--2017
}
