% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{xyt}
\alias{xyt}
\title{Self-exciting point pattern}
\format{
A \code{stppp} object with 653 observations
\describe{
\item{window}{Domain of the point pattern of class \code{owin}}
\item{n}{Number of observations, 653}
\item{x}{x coordinate}
\item{y}{y coordinate}
\item{markformat}{\code{none}}
\item{t}{Timestamp of points}
\item{tlim}{Time frame \code{0 2}}
}
}
\description{
Simulated self-exciting spatiotemporal point pattern of class \code{stppp}
}
